Converted basic files to Typescript
