+++ package.json
@@ -1,19 +1,26 @@
 {
-  "name": "ma-wa",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon ./src/bot.js",
-    "start": "node ./src/bot.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "discord.js": "^12.5.1",
-    "dotenv": "^8.2.0",
-    "fs": "^0.0.1-security"
-  }
+	"name": "ma-wa",
+	"version": "1.0.0",
+	"description": "Axel en Boris zijn cool, rekt",
+	"main": "bot.ts",
+	"scripts": {
+		"test": "echo \"Error: no test specified\" && exit 1",
+		"start": "npm run build:live",
+		"build": "tsc -p .",
+		"build:live": "nodemon --watch 'src/**/*.ts' --exec \"ts-node\" src/bot.ts"
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"dependencies": {
+		"discord.js": "^12.5.1",
+		"dotenv": "^8.2.0",
+		"fs": "^0.0.1-security"
+	},
+	"devDependencies": {
+		"@types/node": "^14.14.35",
+		"nodemon": "^2.0.7",
+		"ts-node": "^9.1.1",
+		"typescript": "^4.2.3"
+	}
 }
