--- src/commands/test/TestCommand.js
@@ -1,11 +1,0 @@
-const BaseCommand = require('../../utils/structures/BaseCommand');
-
-module.exports = class TestCommand extends BaseCommand {
-  constructor() {
-    super('test', 'testing', []);
-  }
-
-  async run(client, message, args) {
-    message.channel.send('Test command works');
-  }
-}
