+++ src/events/message/MessageEvent.ts
@@ -1,0 +1,23 @@
+import BaseEvent from '../../utils/structures/BaseEvent';
+import { Message } from 'discord.js';
+import DiscordClient from '../../client/client';
+
+export default class MessageEvent extends BaseEvent {
+  constructor() {
+    super('message');
+  }
+
+  async run(client: DiscordClient, message: Message) {
+    if (message.author.bot) return;
+    if (message.content.startsWith(client.prefix)) {
+      const [cmdName, ...cmdArgs] = message.content
+        .slice(client.prefix.length)
+        .trim()
+        .split(/\s+/);
+      const command = client.commands.get(cmdName);
+      if (command) {
+        command.run(client, message, cmdArgs);
+      }
+    }
+  }
+}
