+++ src/events/message/MessageReactionRemoveEvent.ts
@@ -1,4 +1,3 @@
-// https://discord.js.org/#/docs/main/stable/class/Client?scrollTo=e-messageReactionRemove
 import { MessageReaction, User } from "discord.js";
 import BaseEvent from "../../utils/structures/BaseEvent";
 import DiscordClient from "../../client/client";
@@ -8,5 +7,25 @@ export default class MessageReactionRemoveEvent extends BaseEvent {
 		super("messageReactionRemove");
 	}
 
-	async run(client: DiscordClient, reaction: MessageReaction, user: User) {}
+	async run(client: DiscordClient, reaction: MessageReaction, user: User) {
+		// try {
+		// 	if (user.bot) return;
+		// 	if (reaction.message.partial) await reaction.message.fetch();
+		// 	if (reaction.partial) await reaction.fetch();
+		// 	if (!reaction.message.guild) return;
+		// 	if (reaction.message.channel.id == process.env.GET_ROLES_CHANNEL_ID) {
+		// 		if (reaction.emoji.name === minecraftEmoji) {
+		// 			return await reaction.message.guild.members.cache
+		// 				.get(user.id)
+		// 				.roles.remove(minecraftRole);
+		// 		} else if (reaction.emoji.name === leagueOfLegendsEmoji) {
+		// 			return await reaction.message.guild.members.cache
+		// 				.get(user.id)
+		// 				.roles.remove(leagueOfLegendsRole);
+		// 		}
+		// 	}
+		// } catch {
+		// 	return;
+		// }
+	}
 }
