+++ src/events/ready/ReadyEvent.ts
@@ -1,0 +1,11 @@
+import BaseEvent from '../../utils/structures/BaseEvent';
+import DiscordClient from '../../client/client';
+
+export default class ReadyEvent extends BaseEvent {
+  constructor() {
+    super('ready');
+  }
+  async run (client: DiscordClient) {
+    console.log('Bot has logged in.');
+  }
+}
