--- src/events/ready/ready.js
@@ -1,10 +1,0 @@
-const BaseEvent = require('../../utils/structures/BaseEvent');
-
-module.exports = class ReadyEvent extends BaseEvent {
-  constructor() {
-    super('ready');
-  }
-  async run (client) {
-    console.log(client.user.tag + ' has logged in.');
-  }
-}
