--- src/utils/structures/BaseCommand.js
@@ -1,7 +1,0 @@
-module.exports = class BaseCommand {
-  constructor(name, category, aliases) {
-    this.name = name;
-    this.category = category;
-    this.aliases = aliases;
-  }
-}
