--- src/utils/structures/BaseEvent.js
@@ -1,5 +1,0 @@
-module.exports = class BaseEvent {
-  constructor(name) {
-    this.name = name;
-  }
-}
